Move errors to top-level expression module
