+++ frontend/src/metabase-lib/v1/expressions/compiler.ts
@@ -2,6 +2,7 @@ import * as Lib from "metabase-lib";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type { Expression } from "metabase-types/api";
 
+import { renderError } from "./errors";
 import { resolverPass } from "./field-resolver";
 import {
   adjustBooleans,
@@ -14,7 +15,6 @@ import {
 } from "./passes";
 import { compile, lexify, parse } from "./pratt";
 import type { ClauseType, ErrorWithMessage, StartRule } from "./types";
-import { renderError } from "./utils";
 
 export type CompileResult<S extends StartRule> =
   | {
