+++ frontend/src/metabase-lib/v1/expressions/resolver.js
@@ -1,8 +1,8 @@
 import { msgid, ngettext, t } from "ttag";
 
 import { MBQL_CLAUSES, getMBQLName } from "./config";
+import { ResolverError } from "./errors";
 import { isCaseOrIfOperator, isOptionsObject } from "./matchers";
-import { ResolverError } from "./pratt/types";
 import { OPERATOR as OP } from "./tokenizer";
 
 const FIELD_MARKERS = ["dimension", "segment", "metric"];
