+++ frontend/src/metabase-lib/v1/expressions/utils.ts
@@ -1,10 +1,6 @@
-import { t } from "ttag";
-
 import * as Lib from "metabase-lib";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
-
-import type { ErrorWithMessage } from "./types";
 
 export function getDatabase(
   query: Lib.Query,
@@ -12,24 +8,6 @@ export function getDatabase(
 ): Database | null {
   const databaseId = Lib.databaseID(query);
   return metadata?.database(databaseId) ?? null;
-}
-
-export function isErrorWithMessage(err: unknown): err is ErrorWithMessage {
-  return (
-    typeof err === "object" &&
-    err != null &&
-    typeof (err as any).message === "string"
-  );
-}
-
-export function renderError(error: unknown) {
-  if (isErrorWithMessage(error) && error.friendly) {
-    return error;
-  }
-  return {
-    message: t`Invalid expression`,
-    friendly: true,
-  };
 }
 
 export function getExpressionMode(startRule: string): Lib.ExpressionMode {
